requirements.txt
torch==2.5.1
torchaudio==2.5.1
transformers==4.44.2
whisperx==3.1.1
pyannote.audio==3.1.1
librosa==0.10.2.post1
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==2.0.2
tqdm==4.66.4
soundfile==0.12.1
streamlit==1.38.0
